requirements.txt
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
langsmith>=0.1.0
streamlit>=1.28.0
openai>=1.3.0